change kiwi to personal fork
This was done because the index(main) file was "index.md" which isn't what I'm using and its not configurable.
So I forked the project and just changed the index file to "README.md" in the code.
The project hasn't been updated in 3 months so it should be fine to use my fork for a long while.